move input prompting code into tui module
